+++ js/main.js
@@ -1,9 +1,10 @@
 import {generateOffers} from './mocks/data.js';
-import {renderOffers} from './similar-offers-list.js';
+import {renderOffer} from './similar-offers-list.js';
 import {disablePage, activatePage} from './form.js';
-const COUNT_OFFER = 1;
 
-renderOffers(generateOffers(COUNT_OFFER));
+const COUNT_OFFER = 10;
+const offers = generateOffers(COUNT_OFFER);
 
+renderOffer(offers[0]);
 disablePage();
 activatePage();
